feat: add image upload and optimization functionality

+++ apps/admin-api/src/modules/subcategory/subcategory.service.ts
@@ -1,10 +1,15 @@
 import { BadRequestException, Injectable } from '@nestjs/common'
 import { prisma } from '@sift-shop/database'
 
+import { optimizationImage } from '~/common/utils/optimizationImage'
+import { S3Service } from '~/infrastructure/s3/s3.service'
+
 import { CreateSubcategoryDto } from './dto/create-subcategory.dto'
 
 @Injectable()
 export class SubcategoryService {
+  constructor(private readonly s3Service: S3Service) {}
+
   async getSubcategories(categoryId?: string) {
     const subcategories = await prisma.subcategory.findMany({
       where: {
@@ -32,7 +37,24 @@ export class SubcategoryService {
     return result
   }
 
-  async createSubcategory(categoryId: string, dto: CreateSubcategoryDto) {
+  async createSubcategory(
+    categoryId: string,
+    file: Express.MulterFile,
+    dto: CreateSubcategoryDto
+  ) {
+    const category = await prisma.category.findFirst({
+      where: {
+        id: categoryId
+      }
+    })
+
+    if (!category) {
+      throw new BadRequestException({
+        code: 'CATEGORY_NOT_FOUND',
+        message: 'Category not found'
+      })
+    }
+
     const subcategory = await prisma.subcategory.findFirst({
       where: {
         OR: [{ name: dto.name }, { slug: dto.slug }]
@@ -46,27 +68,76 @@ export class SubcategoryService {
       })
     }
 
+    const optimizedBuffer = await optimizationImage(file)
+
+    const image = await this.s3Service.uploadFile(
+      optimizedBuffer,
+      'subcategory'
+    )
+
     return await prisma.subcategory.create({
       data: {
         ...dto,
-        categoryId
+        image,
+        categoryId: category.id
       }
     })
   }
 
-  async updateSubcategory(id: string, dto: CreateSubcategoryDto) {
+  async updateSubcategory(
+    id: string,
+    dto: CreateSubcategoryDto,
+    file?: Express.MulterFile
+  ) {
+    const subcategory = await prisma.subcategory.findUnique({
+      where: {
+        id
+      }
+    })
+
+    if (!subcategory) {
+      throw new BadRequestException({
+        code: 'SUBCATEGORY_NOT_FOUND',
+        message: 'Subcategory not found'
+      })
+    }
+
+    let image = subcategory.image
+
+    if (file) {
+      image = await this.s3Service.updateFile(subcategory.image, file)
+    }
+
     return await prisma.subcategory.update({
       where: {
         id
       },
-      data: dto
+      data: {
+        ...dto,
+        image
+      }
     })
   }
 
   async deleteSubcategory(id: string) {
-    return await prisma.subcategory.delete({
+    const subcategory = await prisma.subcategory.findUnique({
       where: {
         id
+      }
+    })
+
+    if (!subcategory) {
+      throw new BadRequestException({
+        code: 'SUBCATEGORY_NOT_FOUND',
+        message: 'Subcategory not found'
+      })
+    }
+
+    void this.s3Service.deleteFile(subcategory.image)
+
+    return await prisma.subcategory.delete({
+      where: {
+        id: subcategory.id
       }
     })
   }
